Remove redundant home repository
